+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - config/mysql.cfg.TEMPLATE:  Point to correct socket file.
 
 ### Changed
+- status:  Processed status return from mongo_libs.ins_doc call.
 - run_program:  Replaced cmds_gen.disconnect with mysql_libs.disconnect call.
 - Removed unnesscary \*\*kwargs arguments.
 - config/mongo.py.TEMPLATE:  Changed configuration entry and added a number of configuration entries.